requirements.txt
pillow
tkinter


